move utilities into utils module
